crimson/net: extract state transition out of repeat_connect()
and extract state transition out of repeat_handle_connect()

in this change, the connect/handle-connect loop is restructured, to
avoid ad-hoc state changes in helper functions. this pave the road to
explicit state transtion using named states.

also, exception is thrown instead in handle_connect_reply(), we should
not proceed in case of failures. and we need do error handling in the
named state in future.

currentl, `state` is set to `state_t::open` in `start_connect()` and
`start_accept()`, the next step is to set it in a named state.

Signed-off-by: Kefu Chai <kchai@redhat.com>