ceph.spec.in: BUILD_WITH_INSTALL_RPATH on SUSE distros
by default, cmake builds with the rpath of build tree, and edit the
rpath when installing the binaries. if the length of the rpath of build
tree is longer than that of the rpath of the installation, cmake stuffs
the rpath with some bogus data. it seems that the linker on openSUSE
leap is more picky at seeing those rpaths and strips them off from the
binaries. cmake fails when trying to change the rpath when installing
binaries, like:

[ 1486s] CMake Error at src/librbd/cmake_install.cmake:70 (file):
[ 1486s]   file RPATH_CHANGE could not write new RPATH:
[ 1486s]
[ 1486s]     /usr/lib64/ceph
[ 1486s]
[ 1486s]   to the file:
[ 1486s]
[ 1486s]     /home/abuild/rpmbuild/BUILDROOT/ceph-16.0.0-3506.ga86c764613a.x86_64/usr/lib64/ceph/librbd/libceph_librbd_parent_cache.so.1.0.0
[ 1486s]
[ 1486s]   No valid ELF RPATH or RUNPATH entry exists in the file;

to avoid this issue, in this change,

CMAKE_BUILD_WITH_INSTALL_RPATH is enabled, when building packages on
SUSE distros to build the packages with the install rpath, so there is no
need to edit the binary anymore.

see also
https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_WITH_INSTALL_RPATH.html
and
http://llvm.1065342.n5.nabble.com/llvm-trunk-build-failed-in-cmake-install-cmake-on-ARM-platform-td65895.html#a65910

Fixes: https://tracker.ceph.com/issues/40295
Signed-off-by: Kefu Chai <kchai@redhat.com>