crimson/os: let get_xattrs() return map<string,bufferptr>
so neither SeaStore::get_attrs() nor AlienStore::get_attrs()
needs to convert the map<string,ptr> to map<string,bufferlist> for
returning it to the client. what client needs is but a encoded
map<string, bufferlist>. but the encoding schemes of buffer::list and
buffer::ptr are identical, so there is no need to convert it back and
forth.

Signed-off-by: Kefu Chai <kchai@redhat.com>