crimson/osd: prepare CrushLocation with do_with()
so we can use it in an async call before sending
"osd crush create-or-move" to monitor. because CrushLocation's
ctor calls the `crush_location_hook`  script in, but we cannot
do this in seastar, otherwise the reactor would be stalled.
as an alternative, we need to

1. futurize `CrushLocation::update_from_hook()`
2. instead of calling it in the ctor, call it in `OSD::_add_me_to_crush`
   explicitly

this change paves the road to the steps listed above.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>