cmake: disable warnings using -Wno-everything
we are not interested in the compiling warning when building
Boost, and even worse we don't want to fail the build because of
the warnings which default to error. and "-w" does not disable
them for us. see also the discussion at
https://reviews.llvm.org/D131307.

to fully disable the warnings, let's add -Wno-everything to the
cflags.

Signed-off-by: Kefu Chai <kchai@redhat.com>